+++ webofneeds/won-core/src/main/java/won/protocol/repository/DatasetHolderRepository.java
@@ -26,7 +26,7 @@ import java.net.URI;
  */
 public interface DatasetHolderRepository extends CrudRepository<DatasetHolder, URI>
 {
-  public DatasetHolder findOneByUriAndVersionNot(URI uri, Long version);
+  public DatasetHolder findOneByUriAndVersionNot(URI uri, Integer version);
 
   public DatasetHolder findOneByUri(URI uri);
 }
